ref #4 Terminus value checks now work by exceptions.
Now the Terminus derived check classes should raise type and value
errors with a short message. This is caught by the Terminus class which
raises a CheckError. The settings derived class which is trying to
instantiate the Terminus instance catches this CheckError and raises an
AssignError. The error message will print the name of the setting, the
passed value, and why it failed the check.