ref #3 Added a new settings file called Dict.
Ive added an equivalent class to List but for dictionaries. Now you can
have a settings class derive from Dict, pass a dictionary to its
constructor and it will create an object which behaves like a dict but
has the restriction that all values are of the same type, either
primitive or other settings derived objects.