ref #17 Traceable error functionality now exists.
Modifications to the way the Settings derived classes (and Terminus
derived) raise and catch exceptions, and the nature of the exceptions
have been made. When a SettingTypeError or SettingCheckError is raised a
SettingErrorMessage error is raised recusively up to the base of the
Settings 'tree'. Each successive SettingsErrorMessage contains
information about other that were raised before it.

Dict and List are now derived from Settings to reduce code redundancy.

Dict, List, and Terminus no longer have a restriction to a single
attribute.

Dict, Terminus, and List derived classes must have
self.type = <some type> in the child __init__ method.