ref #16 Added new Number class with optional range
The new number class is a new type of base settings class that allows
numerical settings to have a range of values specified. It can only be
an terminating setting (becuase its a number). The child class specifies
a self.type = <type> like the Dict, Terminus and List classes, but its
constructor can take a primitive of type <type>, a List[<type>] or a
dictionary {"min": <type>, "max": <type>, "num": <int>}.