ref #16 Updated IDMFConfig with new number class
The new Number base class has been implemented in the idmf config file.
Now any parameters which were numberical can be provided as a list of
values or a range, or as a single value.

Methods have been added for upper and lower bound checks. These can be
called in the derived class implemenetation of the check method.