ref #15 Added a ComputationalSpaceBuilder class
In the process of implementing the run containers, a computational space
builder class was created. This class takes a Settings derived object and
returns creates a dictionary of the the same Settings derived object
where each new instance is a sample from the configuration space defined
by the parameters which had range definitions in the original settings
object.