refs 13 Bug Fixes and Structural Changes.
Changed the release type models to have the single underscore instead of the double as getattr() could not find the private method. Fixed a bug in fixed release where the concentration was exponentially increasing when no sources were emitting. Removed c_end and t_end as we already know that c_end is the source.end_time and we can work out c_end from that index value.