refs #20 Added run method for plots.
Implemented the first rendition of the run method. This calculated the concentration given the config, then plots and animated the points, lines and planes given in the config. Additionally there are point, line and plane extractors which only plot the section of concentration which is being monitored. Error with the Line animation causing overlapping lines each plot.