ref #38 CSVToConfigFile class implemented.
The CSVToConfigFile class takes a csv file where each row starts with
either INS, INF, FIX, POI, LIN, PLA and then some number of parameters
which depends on  the first entry in the row. Each line corresponds to a
source or monitor location. The config file that is passed to the
constructor is read and the entries specified in the CSV file are added.
The new config file is written to disk with the '_new' prefix or a
given file name.